services/run_routine/requirements.txt
blinker==1.9.0
click==8.2.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
paho-mqtt==2.1.0
psycopg2-binary==2.9.10
python-dotenv==1.1.0
Werkzeug==3.1.3
